+++ modules/c-api/src/CnaCApiGraphics.cpp
@@ -283,6 +283,8 @@ static_assert(
     static_cast<uint32_t>(CNA::RendererFeature::ShaderDialectWgsl) ==
         CNA_RENDERER_FEATURE_SHADER_DIALECT_WGSL &&
     static_cast<uint32_t>(CNA::RendererFeature::Texture3DSampling) ==
+        CNA_RENDERER_FEATURE_TEXTURE_3D_SAMPLING &&
+    static_cast<uint32_t>(CNA::RendererFeature::BaseInstanceDrawing) ==
         CNA_RENDERER_FEATURE_MAXIMUM &&
     static_cast<uint32_t>(CNA::RendererFeature::Count) ==
         CNA_RENDERER_FEATURE_MAXIMUM + UINT32_C(1));